refactor: util 클래스 구분
