+++ src/main/java/racingcar/Game.java
@@ -2,6 +2,7 @@ package racingcar;
 
 import java.util.List;
 import racingcar.car.Cars;
+import racingcar.utils.NameExtractor;
 
 public class Game {
