+++ src/main/java/racingcar/car/Constants.java
@@ -4,6 +4,7 @@ public final class Constants {
 
     public static final String EMPTY_STRING = "";
     public static final int MAX_LENGTH_OF_NAME = 5;
+    public static final int MOVEABLE_MIN_NUMBER = 4;
 
     private Constants() {
     }
