+++ src/main/java/racingcar/utils/NameExtractor.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.utils;
 
 import java.util.ArrayList;
 import java.util.List;
